requirements.txt
fastapi>=0.129.0
uvicorn[standard]>=0.40.0
pydantic>=2.12.0
pyswisseph==2.10.3.2
pyyaml>=6.0.2
timezonefinder>=6.5.0
python-multipart>=0.0.20
